:recycle: Move animations into Resources
